feat: point tracker working, unscuffed
