+++ points.json
@@ -1,1 +1,1 @@
-{"fresco#0001":{"level":1,"needed":10,"points":16}}
+{"fresco#0001":{"level":3,"needed":29,"points":17},"fresco#4852":{"level":2,"needed":17,"points":10}}
